+++ src/color/colormaker-registry.js
@@ -194,16 +194,20 @@ class ColormakerRegistry{
 
     }
 
-    addSelectionScheme( pairList, label ){
+    addSelectionScheme( dataList, label ){
 
         class MySelectionColormaker extends SelectionColormaker{
             constructor( params ){
-                super( Object.assign( { pairList: pairList }, params ) );
+                super( Object.assign( { dataList: dataList }, params ) );
             }
         }
 
         return this.addUserScheme( MySelectionColormaker, label );
 
+    }
+
+    hasScheme( id ) {
+        return id in this.schemes || id in this.userSchemes;
     }
 
 }