+++ src/color/selection-colormaker.js
@@ -9,6 +9,7 @@ import { Color } from "../../lib/three.es6.js";
 
 import Selection from "../selection.js";
 import Colormaker from "./colormaker.js";
+import { ColormakerRegistry } from "../globals.js";
 
 
 class SelectionColormaker extends Colormaker{
@@ -17,25 +18,44 @@ class SelectionColormaker extends Colormaker{
 
         super( params );
 
-        this.pairList = params.pairList || [];
+        const dataList = params.dataList || [];
 
-        this.colorList = [];
+        this.colormakerList = [];
         this.selectionList = [];
 
-        this.pairList.forEach( pair => {
-            this.colorList.push( new Color( pair[ 0 ] ).getHex() );
-            this.selectionList.push( new Selection( pair[ 1 ] ) );
+        dataList.forEach( pair => {
+
+            const [ scheme, sele, params={} ] = pair;
+
+            if( ColormakerRegistry.hasScheme( scheme ) ){
+                Object.assign( params, {
+                    scheme: scheme,
+                    structure: this.structure
+                } );
+            }else{
+                Object.assign( params, {
+                    scheme: "uniform",
+                    value: new Color( scheme ).getHex()
+                } );
+            }
+
+            this.colormakerList.push( ColormakerRegistry.getScheme( params ) );
+            this.selectionList.push( new Selection( sele ) );
+
         } );
 
     }
 
     atomColor( a ){
-        for( var i = 0, n = this.pairList.length; i < n; ++i ){
+
+        for( let i = 0, n = this.selectionList.length; i < n; ++i ){
             if( this.selectionList[ i ].test( a ) ){
-                return this.colorList[ i ];
+                return this.colormakerList[ i ].atomColor( a );
             }
         }
+
         return 0xFFFFFF;
+
     }
 
 }