+++ src/ngl.js
@@ -116,6 +116,7 @@ import "./parser/dcd-parser.js";
 
 import "./parser/mrc-parser.js";
 import "./parser/cube-parser.js";
+import "./parser/dsn6-parser.js";
 import "./parser/dx-parser.js";
 import "./parser/dxbin-parser.js";
 import "./parser/xplor-parser.js";